requirements.txt
streamlit>=1.32.0
google-generativeai>=0.8.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
streamlit-autorefresh>=1.0.1
